move config location to sub folder
