+++ dbt_platform_helper/entities/platform/extensions/opensearch.py
